Split done, ready to test
